centralize implementation in docker-compose-rule-core
