+++ docker-compose-rule-core/src/integrationTest/java/com/palantir/docker/compose/RemoveConflictingContainersIntegrationTest.java
@@ -33,11 +33,11 @@ public class RemoveConflictingContainersIntegrationTest {
 
     @Test
     public void removeConflictingContainersOnStartup_off_should_fail_fast_if_containers_exist() throws IOException, InterruptedException {
-        DockerComposeRule composition = DockerComposeRule.builder()
+        DockerComposeManager composition = DockerComposeManager.testBuilder()
                 .file(DOCKER_COMPOSE_YAML_PATH)
                 .retryAttempts(0)
                 .build();
-        DockerComposeRule conflictingComposition = DockerComposeRule.builder()
+        DockerComposeManager conflictingComposition = DockerComposeManager.testBuilder()
                 .file(DOCKER_COMPOSE_YAML_PATH)
                 .retryAttempts(0)
                 .removeConflictingContainersOnStartup(false)
@@ -55,7 +55,7 @@ public class RemoveConflictingContainersIntegrationTest {
 
     @Test
     public void by_default_existing_containers_should_be_removed_silently() throws IOException, InterruptedException {
-        DockerComposeRule composition = DockerComposeRule.builder()
+        DockerComposeManager composition = DockerComposeManager.testBuilder()
                 .file(DOCKER_COMPOSE_YAML_PATH)
                 .retryAttempts(0)
                 .shutdownStrategy(AGGRESSIVE)
