+++ docker-compose-rule-core/src/integrationTest/java/com/palantir/docker/compose/execution/EnvironmentVariableIntegrationTest.java
@@ -20,7 +20,7 @@ import static com.palantir.docker.compose.matchers.IOMatchers.fileContainingStri
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 
-import com.palantir.docker.compose.DockerComposeRule;
+import com.palantir.docker.compose.DockerComposeManager;
 import com.palantir.docker.compose.connection.DockerMachine;
 import java.nio.file.Path;
 import org.junit.Rule;
@@ -38,7 +38,7 @@ public class EnvironmentVariableIntegrationTest {
                                                    .withAdditionalEnvironmentVariable("SOME_VARIABLE", "SOME_VALUE")
                                                    .build();
 
-        DockerComposeRule dockerComposition = DockerComposeRule.builder()
+        DockerComposeManager dockerComposition = DockerComposeManager.testBuilder()
                 .file("src/test/resources/environment/docker-compose.yaml")
                 .machine(dockerMachine)
                 .waitingForService("env-test", toHaveAllPortsOpen())
