+++ docker-compose-rule-core/src/main/java/com/palantir/docker/compose/RecordingCluster.java
