+++ docker-compose-rule-core/src/test/resources/docker-compose.yaml
@@ -1,0 +1,23 @@
+db:
+  image: appropriate/nc
+  command: /bin/sh -c 'echo server started && nc -lk 5432'
+  ports:
+    - "5432"
+
+db2:
+  image: appropriate/nc
+  command: /bin/sh -c 'echo server started && nc -lk 5432'
+  ports:
+    - "5432"
+
+db3:
+  image: appropriate/nc
+  command: /bin/sh -c 'echo server started && nc -lk 5432'
+  ports:
+    - "5432"
+
+db4:
+  image: appropriate/nc
+  command: /bin/sh -c 'echo server started && nc -lk 5432'
+  ports:
+    - "5432"
