+++ docker-compose-rule-core/src/test/resources/environment/docker-compose.yaml
