+++ docker-compose-rule-core/src/test/resources/events/all-good-docker-compose.yaml
