+++ docker-compose-rule-core/src/test/resources/host-networked-docker-compose.yaml
