+++ docker-compose-rule-core/src/test/resources/named-containers-docker-compose.yaml
