+++ docker-compose-rule-core/src/test/resources/shutdown-strategy-with-network.yaml
