+++ docker-compose-rule-core/src/test/resources/shutdown-strategy.yaml
