+++ docker-compose-rule-core/src/test/resources/up-integration-test.yaml
