--- docker-compose-rule-junit4/src/test/resources/native-healthcheck.yaml
@@ -1,11 +1,0 @@
-version: "2.1"
-
-services:
-  dummy:
-    image: gliderlabs/alpine:3.4
-    command: sh -c 'while true; do sleep 10; done'
-    healthcheck:
-      test: ["CMD", "test", "-f", "healthy"]
-      interval: 100ms
-      timeout: 1s
-      retries: 1
